refactor: split Pom and Dependency classes into separate files for better organization
