+++ src/main/java/org/buildcli/utils/Pom.java
@@ -108,75 +108,3 @@ public class Pom {
         return sb.toString();
     }
 }
-
-class Dependency{
-    private String groupId;
-    private String artifactId;
-    private String version;
-    private String type;
-    private String scope;
-    private String optional;
-
-    public Dependency(String groupId, String artifactId, String version) {
-        this.groupId = groupId;
-        this.artifactId = artifactId;
-        this.version = version;
-    }
-
-    public Dependency(String groupId, String artifactId, String version, String type, String scope, String optional) {
-        this.groupId = groupId;
-        this.artifactId = artifactId;
-        this.version = version;
-        this.type = type;
-        this.scope = scope;
-        this.optional = optional;
-    }
-
-    public String getGroupId() {
-        return groupId;
-    }
-
-    public void setGroupId(String groupId) {
-        this.groupId = groupId;
-    }
-
-    public String getArtifactId() {
-        return artifactId;
-    }
-
-    public void setArtifactId(String artifactId) {
-        this.artifactId = artifactId;
-    }
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getScope() {
-        return scope;
-    }
-
-    public void setScope(String scope) {
-        this.scope = scope;
-    }
-
-    public String getOptional() {
-        return optional;
-    }
-
-    public void setOptional(String optional) {
-        this.optional = optional;
-    }
-}
